fixup! Move `airflow.utils.operator_resources` to task-sdk
