+++ task-sdk/src/airflow/sdk/bases/operator.py
@@ -77,10 +77,10 @@ if TYPE_CHECKING:
 
     import jinja2
 
-    from airflow.sdk.bases.operator_resources import Resources
     from airflow.sdk.bases.operatorlink import BaseOperatorLink
     from airflow.sdk.definitions.context import Context
     from airflow.sdk.definitions.dag import DAG
+    from airflow.sdk.definitions.operator_resources import Resources
     from airflow.sdk.definitions.taskgroup import TaskGroup
     from airflow.sdk.definitions.xcom_arg import XComArg
     from airflow.serialization.enums import DagAttributeTypes
@@ -177,7 +177,7 @@ def coerce_timedelta(value: float | timedelta, *, key: str | None = None) -> tim
 def coerce_resources(resources: dict[str, Any] | None) -> Resources | None:
     if resources is None:
         return None
-    from airflow.sdk.bases.operator_resources import Resources
+    from airflow.sdk.definitions.operator_resources import Resources
 
     return Resources(**resources)
 
@@ -1330,7 +1330,7 @@ class BaseOperator(AbstractOperator, metaclass=BaseOperatorMeta):
         if resources is None:
             return None
 
-        from airflow.sdk.bases.operator_resources import Resources
+        from airflow.sdk.definitions.operator_resources import Resources
 
         if isinstance(resources, Resources):
             return resources
