+++ task-sdk/src/airflow/sdk/definitions/operator_resources.py
